feat: @fly/fetch/mount, easily mount paths
The `mount` function takes a map of path prefixes => `fetch` functions,
and checks the pathname on incoming requests, then routes to the
appropriate `fetch` function.
