+++ src/utils/build.ts
@@ -106,9 +106,11 @@ export function getWebpackConfig(cwd: string, opts?: AppBuilderOptions): webpack
       "@fly/proxy": v8EnvPath + "/fly/proxy",
       "@fly/data": v8EnvPath + "/fly/data",
       "@fly/cache": v8EnvPath + "/fly/cache",
-      "@fly/static": v8EnvPath + "/fly/static"
+      "@fly/static": v8EnvPath + "/fly/static",
+      "@fly/fetch": v8EnvPath + "/fly/fetch"
     })
   }, conf.resolve)
+
 
   if (opts && opts.uglify) {
     conf.plugins = conf.plugins || []
